ToontownGUI: Fix logo scale calculation
IDEK what I was thinking whenI wrote the original stuff, but now it
makes sense. The aspect ratio of this image is 2:1, so we simply multiply
the scale we want by 2 on the X, and leave it alone on the Z (or for
nerds who hate Panda3D, te Y IDK math is stupid, which is why I usually
leave it to Peter).

Lord save me.